refactor(community): tiptap-content-section 컴포넌트화
