+++ apps/community/src/shared/components/tiptap-content-section/tiptap-content-section.css.ts
